+++ R/last_fit.R
@@ -1,0 +1,6 @@
+
+#' @export
+orbital.last_fit <- function(x, ...) {
+  x <- hardhat::extract_workflow(x)
+  orbital(x)
+}